+++ src/main/java/com/example/practice_modsen_shop/dto/OrderItemRequestTo.java
@@ -1,4 +1,5 @@
-package com.example.practice_modsen_shop.DTO;
+package com.example.practice_modsen_shop.dto;
+
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -7,8 +8,7 @@ import lombok.Setter;
 @NoArgsConstructor
 @Getter
 @Setter
-public class OrderItemResponseTo {
-
+public class OrderItemRequestTo {
     private Long id;
 
     private Long productId;
